requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
db-dtypes==1.5.0
google-api-core==2.30.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.78.1
idna==3.11
numpy==2.0.2
oauthlib==3.3.1
packaging==26.0
pandas==2.3.3
pandas-gbq==0.33.0
pip==22.0.4
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==2.23
pydata-google-auth==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==58.1.0
six==1.17.0
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uv==0.10.4
